hotfix: put out a fire that disappear bingbong code
