+++ frontend/src/components/_common/Header/Header.styles.ts
@@ -20,8 +20,6 @@ export const s_logoContainer = css`
   display: flex;
   gap: 1.2rem;
   align-items: center;
-
-  width: 100%;
   height: 3.6rem;
 `;
