+++ frontend/src/mocks/meeting/meetingHandlers.ts
@@ -3,8 +3,8 @@ import { HttpResponse, http } from 'msw';
 import { BASE_URL } from '@constants/api';
 
 import meetingAllSchedules from './data/meetingAllSchedules.json';
+import meetingSingleSchedule from './data/meetingSingleSchedule.json';
 import meetingTableFrame from './data/meetingTableFrame.json';
-import oneAttendeeSchedules from './data/oneAttendeeSchedule.json';
 
 const meetingHandlers = [
   http.get(`${BASE_URL}/:uuid`, () => {
@@ -18,7 +18,7 @@ const meetingHandlers = [
     if (attendeeName === 'all') {
       return HttpResponse.json(meetingAllSchedules, { status: 200 });
     }
-    return HttpResponse.json(oneAttendeeSchedules, { status: 200 });
+    return HttpResponse.json(meetingSingleSchedule, { status: 200 });
   }),
 ];
